# docs/openapi.yaml
openapi: 3.0.0
info:
  version: 0.1.0
  title: SplitNaira API
  description: Premium royalty management API on Stellar network.
servers:
  - url: http://localhost:3001
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from POST /users/login
    adminApiKey:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Payments admin API key for /splits/admin mutation endpoints
  schemas:
    ApiError:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        requestId:
          type: string
          description: Request correlation ID
        details:
          type: object
          additionalProperties:
            nullable: true
          description: Additional error context
      required:
        - error
      example:
        error: VALIDATION_ERROR
        message: collaborators basisPoints must sum to exactly 10000
        requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
        details:
          path:
            - collaborators
...
paths:
...
  /events:
    get:
      summary: Observe transaction status updates for a submitted XDR (SSE)
      description: |-
        Opens a Server-Sent Events stream and emits a `transaction_update` event each time the backend observes an update for the given `txHash`. Limited to 5 concurrent subscribers per `txHash` (configurable via `SSE_MAX_LISTENERS_PER_TXHASH`); exceeding it returns 429.

        Example frame:
        ```
        event: transaction_update
        data: {"txHash":"3389e9f0f1a65f19736cacf544c2e825313e8447f569233bb8db39aa607c8fc","status":"pending"}

        ```
      tags:
        - Events
      parameters:
        - schema:
            type: string
            minLength: 1
            example: 3389e9f0f1a65f19736cacf544c2e825313e8447f569233bb8db39aa607c8fc
            description: Transaction hash to observe
          required: true
          description: Transaction hash to observe
          name: txHash
          in: query
      responses:
        "200":
          description: text/event-stream of `transaction_update` frames (connection stays
            open)
          content:
            text/event-stream:
              schema:
                type: string
                example: |+
                  event: transaction_update
                  data: {"txHash":"3389e9f0f1a65f19736cacf544c2e825313e8447f569233bb8db39aa607c8fc","status":"pending"}

        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Too many concurrent subscribers for this txHash
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: too_many_event_listeners
                code: RESOURCE_LIMIT_EXCEEDED
                message: Too many event stream subscribers for this transaction.
                requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
                details:
                  txHash: 3389e9f0f1a65f19736cacf544c2e825313e8447f569233bb8db39aa607c8fc
                  limit: 5
  /events/transactions/{txHash}:
    get:
      summary: Observe a single transaction until it's confirmed (SSE)
      description: |-
        Opens a Server-Sent Events stream and emits one `transaction:confirmed` event with the full transaction record once `EventListenerService` observes it on-chain. Sends a `: heartbeat` comment every 15s to keep intermediary proxies from closing the connection.

        Example frame:
        ```
        event: transaction:confirmed
        data: {"txHash":"3389e9f0f1a65f19736cacf544c2e825313e8447f569233bb8db39aa607c8fc","roundId":"afrobeats_001","recipient":"GCZTK3PNYXCVF7Z3ELVELEJZ7BVGPGPVJ3NVZNI7O6DVCHPA4JU6LXRV","amount":"1000","token":"native","timestamp":1732012800,"status":"completed"}

        ```
      tags:
        - Events
      parameters:
        - schema:
            type: string
            minLength: 1
            example: 3389e9f0f1a65f19736cacf544c2e825313e8447f569233bb8db39aa607c8fc
          required: true
          name: txHash
          in: path
      responses:
        "200":
          description: text/event-stream of a single `transaction:confirmed` frame
            (connection stays open until then)
          content:
            text/event-stream:
              schema:
                type: string
                example: |+
                  event: transaction:confirmed
                  data: {"txHash":"3389e9f0f1a65f19736cacf544c2e825313e8447f569233bb8db39aa607c8fc","roundId":"afrobeats_001","recipient":"GCZTK3PNYXCVF7Z3ELVELEJZ7BVGPGPVJ3NVZNI7O6DVCHPA4JU6LXRV","amount":"1000","token":"native","timestamp":1732012800,"status":"completed"}

        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /transactions/{txHash}:
    get:
      summary: Get transaction by hash
      description: Returns a single payout record matching the Stellar transaction hash.
      tags:
        - Transactions
      parameters:
        - schema:
            type: string
            minLength: 1
          required: true
          name: txHash
          in: path
      responses:
        "200":
          description: Transaction record
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  roundId:
                    type: string
                  recipient:
                    type: string
                  amount:
                    type: string
                  token:
                    type: string
                  timestamp:
                    type: integer
                  txHash:
                    type: string
                  status:
                    type: string
                    enum:
                      - pending
                      - completed
                      - failed
                required:
                  - id
                  - roundId
                  - recipient
                  - amount
                  - token
                  - timestamp
                  - txHash
                  - status
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: NOT_FOUND
                message: Project afrobeats_001 does not exist.
                requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: INTERNAL_ERROR
                message: An unexpected error occurred. Please try again later.
                requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
...